fix seeds file and update
